rename: DTO 디렉토리 분리 및 DTO명 수정 - DTO가 늘어남에 따라 dto 디렉토리 내에서 request, response 디렉토리를 분리했습니다. - DTO명 중복이 예상되어 DTO명을 더 명시적으로 변경하였습니다.
