+++ src/main/java/com/example/muaring/domain/member/service/MemberService.java
@@ -10,6 +10,7 @@ import com.example.muaring.domain.file.repository.ImageRepository;
 import com.example.muaring.domain.file.service.ImageService;
 import com.example.muaring.domain.member.dto.request.MemberProfileCreateRequestDTO;
 import com.example.muaring.domain.member.dto.response.MemberProfileResponseDTO;
+import com.example.muaring.domain.member.dto.response.MemberProfileCreateResponseDTO;
 import com.example.muaring.domain.member.dto.response.NicknameCheckResponseDTO;
 import com.example.muaring.domain.member.entity.Member;
 import com.example.muaring.domain.file.entity.Image;
@@ -39,7 +40,7 @@ public class MemberService {
     }
 
     @Transactional
-    public MemberProfileResponseDTO registerProfile(MemberProfileCreateRequestDTO requestDTO) {
+    public MemberProfileCreateResponseDTO registerProfile(MemberProfileCreateRequestDTO requestDTO) {
         String s3Key = requestDTO.s3Key();
 
         try {
@@ -51,10 +52,15 @@ public class MemberService {
                 throw new MemberException(MemberErrorCode.ALREADY_HAS_PROFILE);
             }
 
-            Image image = createImage(requestDTO);
+            Image image = createImage(
+                    requestDTO.fileName(),
+                    requestDTO.fileType(),
+                    requestDTO.imageType(),
+                    requestDTO.s3Key(),
+                    requestDTO.fileSize());
 
-            member.updateProfile(requestDTO.nickname(), image);
-            return MemberProfileResponseDTO.of(member, s3Properties.s3().bucketUrl(s3Properties.region()));
+            member.createProfile(requestDTO.nickname(), image);
+            return MemberProfileCreateResponseDTO.of(member, s3Properties.s3().bucketUrl(s3Properties.region()));
         } catch (Exception e) {
             if (s3Key != null && !s3Key.isBlank()) {
                 log.warn("❌ 프로필 정보 생성 도중 문제가 발생했습니다.");
@@ -64,9 +70,9 @@ public class MemberService {
         }
     }
 
-    private Image createImage(MemberProfileCreateRequestDTO requestDTO) {
+    private Image createImage(String fileName, String fileType, ImageType imageType, String s3Key, Long fileSize) {
         // 이미지를 선택하지 않은 경우 기본 이미지 반환
-        if (requestDTO.s3Key() == null || requestDTO.s3Key().isEmpty()) {
+        if (s3Key == null || s3Key.isEmpty()) {
             ImageProperties.DefaultImageProperties defaultProfileImage = imageProperties.defaultProfile();
             return imageRepository.findByImageTypeAndS3Key(ImageType.MEMBER, defaultProfileImage.s3Key())
                     .orElseThrow(() -> new FileException(FileErrorCode.IMAGE_NOT_FOUND));
@@ -74,11 +80,11 @@ public class MemberService {
 
         // 이미지를 선택한 경우 presigned 업로드 이미지 메타데이터 생성
         Image image = Image.create(
-                requestDTO.fileName(),
-                requestDTO.fileType(),
-                ImageType.MEMBER,
-                requestDTO.s3Key(),
-                requestDTO.fileSize()
+                fileName,
+                fileType,
+                imageType,
+                s3Key,
+                fileSize
         );
         return imageRepository.save(image);
     }
